requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.14
pyautogui==0.9.54
psutil==5.9.5
requests==2.31.0
opencv-python==4.8.1
Pillow==10.0.0
python-dotenv==1.0.0
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
spacy==3.7.2
joblib==1.3.2
google-cloud-texttospeech==2.14.1
transformers==4.37.2
torch==2.2.0
